+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.1.1 - 2023/03/18
+
+### Added
+- More Teddit API endpoints
+
+### Fixed
+- Redgifs videos
+
 # 2.1.0 - 2022/12/28
 
 ### Added